@@ -1,17 +1,17 @@
 source 'https://rubygems.org'
-ruby '2.2.0'
-gem 'rails', '4.2.5'
+ruby '2.3.0'
+gem 'rails', '5.0.0'
 gem 'pg'
-gem 'sass-rails', '~> 5.0'
+gem 'sass-rails', '>= 5.0.6'
 gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.1.0'
-gem 'jquery-rails'
-gem 'jbuilder', '~> 2.0'
-gem 'annotate'
-
-gem 'activeadmin', github: 'activeadmin'
+gem 'coffee-rails', '~> 4.2.1'
+gem 'jquery-rails', '>= 4.2.1'
+gem 'jbuilder', '>= 2.6.0'
+gem 'annotate', '>= 2.7.1'
+gem 'inherited_resources', github: 'activeadmin/inherited_resources'
+gem 'activeadmin', '~> 1.0.0.pre4' #github: 'activeadmin'
 gem 'paper_trail', '~> 4.0.0'
-gem 'dotenv-rails'
+gem 'dotenv-rails', '>= 2.1.1'
 # MailChimp integration
 gem 'gibbon'
 # The push notification service.
@@ -20,28 +20,31 @@ group :development, :test do
   gem 'byebug'
 end
 group :development do
-  gem 'web-console', '~> 2.0'
+  #gem 'web-console', '~> 2.0' # Dont and caused Rails 5 problem
   gem 'spring'
   gem 'letter_opener'
 end
-gem 'apipie-rails'
+#gem 'apipie-rails'
 gem 'devise'
 gem 'high_voltage'
-gem 'simple_form'
+gem 'simple_form','>= 3.3.1'
 gem 'rack-cors', :require => 'rack/cors'
-gem 'omniauth'
-gem 'devise_token_auth'
-gem 'sidekiq', '~> 4.0', '>= 4.0.1'
+gem 'omniauth', '>= 1.3.1'
+gem 'devise_token_auth', '>= 0.1.39'
+gem 'sidekiq', '>= 4.2.2'
 gem 'ckeditor', '~> 4.1', '>= 4.1.5'
 gem 'socket.io-client-simple'
 gem 'houston'
 gem 'fancybox-rails', '~> 0.3.0'
+
 gem 'mini_magick', '~> 4.3', '>= 4.3.6'
 gem 'carrierwave', '~> 0.10.0'
 gem 'fog', '~> 1.36'
+
+#gem 'responders'
 group :development do
   gem 'better_errors'
-  gem 'quiet_assets'
+  #gem 'quiet_assets','>= 1.1.0' # No compatible version
   gem 'rails_layout'
   gem 'spring-commands-rspec'
   gem 'capistrano', '~> 3.1.0'
@@ -50,13 +53,14 @@ group :development do
   gem 'capistrano-rvm', github: "capistrano/rvm"
   gem 'capistrano-sidekiq'
   gem 'rest-client'
+  gem 'puma'
 end
 group :development, :test do
   gem 'factory_girl_rails'
   gem 'faker'
-  gem 'rspec'
+  #gem 'rspec'
   gem 'fakeredis'
-  gem 'rspec-rails'
+  gem 'rspec-rails', '~> 3.5.2'
   gem 'shoulda-matchers', '~> 3.0'
   gem 'shoulda-callback-matchers', '~> 1.1.1'
   gem 'redis-namespace'
@@ -67,4 +71,4 @@ group :test do
   gem 'launchy'
   gem 'selenium-webdriver'
 end
-gem 'sinatra', :require => nil
+gem 'sinatra', :require => nil #may re-enable it
