@@ -2,7 +2,7 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead
   #protect_from_forgery with: :exception
-  before_filter :log_user_sign_out, if: -> { request.url.include?'/sign_out' }
+  before_action :log_user_sign_out, if: -> { request.url.include?'/sign_out' }
 
   def authenticate_active_admin_user!
     authenticate_user!
