@@ -36,7 +36,7 @@
 #  device_token           :string
 #  device_type            :string
 
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   # Include default devise modules.
   devise :database_authenticatable, :registerable,
           :recoverable, :rememberable, :trackable, :validatable,
