@@ -1,15 +1,6 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
-require "rails"
-# Pick the frameworks you want:
-require "active_model/railtie"
-require "active_job/railtie"
-require "active_record/railtie"
-require "action_controller/railtie"
-require "action_mailer/railtie"
-require "action_view/railtie"
-require "sprockets/railtie"
-# require "rails/test_unit/railtie"
+require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -17,46 +8,37 @@ Bundler.require(*Rails.groups)
 
 module Fram
   class Application < Rails::Application
-
-    config.generators do |g|
-      g.helper false
-      g.javascript_engine false
-      g.stylesheets false
-      g.test_framework :rspec,
-        fixtures: true,
-        view_specs: false,
-        helper_specs: false,
-        routing_specs: false,
-        controller_specs: true,
-        request_specs: false
-      g.fixture_replacement :factory_girl, dir: "spec/factories"
-    end
+    #config.generators do |g|
+    #  g.helper false
+    #  g.javascript_engine false
+    #  g.stylesheets false
+    #  g.test_framework :rspec,
+    #    fixtures: true,
+    #    view_specs: false,
+    #    helper_specs: false,
+    #    routing_specs: false,
+    #    controller_specs: true,
+    #    request_specs: false
+    #  g.fixture_replacement :factory_girl, dir: "spec/factories"
+    #end
     
-    config.middleware.use Rack::Cors do
-      allow do
-        origins '*'
-        resource '*',
-          :headers => :any,
-          :expose  => ['access-token', 'expiry', 'token-type', 'uid', 'client'],
-          :methods => [:get, :post, :options, :delete, :put]
-      end
-    end
+    #config.middleware.use Rack::Cors do
+    #  allow do
+    #    origins '*'
+    #    resource '*',
+    #      :headers => :any,
+    #      :expose  => ['access-token', 'expiry', 'token-type', 'uid', 'client'],
+    #      :methods => [:get, :post, :options, :delete, :put]
+    #  end
+    #end
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
-
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
-
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
     config.assets.precompile += ['active_admin.css', 'active_admin.js']
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    #config.active_record.raise_in_transactional_callbacks = true
     # Setup sidkiq to be used for delayed jobs
-    config.active_job.queue_adapter = :sidekiq
+    #config.active_job.queue_adapter = :sidekiq
     # Setup autoload for lib
     config.autoload_paths += Dir["#{Rails.root}/lib"]
   end
