@@ -1,7 +1,7 @@
-Apipie.configure do |config|
-  config.app_name                = "Framework"
-  config.api_base_url            = "/api/v1"
-  config.doc_base_url            = "/apipie"
+#Apipie.configure do |config|
+#  config.app_name                = "Framework"
+#  config.api_base_url            = "/api/v1"
+#  config.doc_base_url            = "/apipie"
   # where is your API defined?
-  config.api_controllers_matcher = "#{Rails.root}/app/controllers/api/v1/*.rb"
-end
+#  config.api_controllers_matcher = "#{Rails.root}/app/controllers/api/v1/*.rb"
+#end
